@@ -7,7 +7,7 @@
     {
       "description": "Jest Coverage",
       "href": "https://github.com/mflorence99/lintel/issues",
-      "url": "https://img.shields.io/badge/coverage-97.66%25-green"
+      "url": "https://img.shields.io/badge/coverage-97.67%25-green"
     },
     {
       "description": "License: MIT",
@@ -57,7 +57,10 @@
       "title": "Lintel"
     }
   },
-  "dependencies": {},
+  "dependencies": {
+    "eslint": "^7",
+    "js-yaml": "^3.14.0"
+  },
   "description": "VSCode extension to visualize ESLint configurations",
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1000.0",
@@ -66,15 +69,15 @@
     "@angular-eslint/schematics": "0.0.1-alpha.32",
     "@angular-eslint/template-parser": "0.0.1-alpha.32",
     "@angular/cdk": "^10",
-    "@angular/cli": "~10",
-    "@angular/common": "~10",
-    "@angular/compiler": "~10",
-    "@angular/compiler-cli": "~10",
-    "@angular/core": "~10",
-    "@angular/forms": "~10",
-    "@angular/language-service": "~10",
-    "@angular/platform-browser": "~10",
-    "@angular/platform-browser-dynamic": "~10",
+    "@angular/cli": "^10",
+    "@angular/common": "^10.0.1",
+    "@angular/compiler": "^10.0.1",
+    "@angular/compiler-cli": "^10.0.1",
+    "@angular/core": "^10.0.1",
+    "@angular/forms": "^10.0.1",
+    "@angular/language-service": "^10.0.1",
+    "@angular/platform-browser": "^10.0.1",
+    "@angular/platform-browser-dynamic": "^10.0.1",
     "@fortawesome/angular-fontawesome": "^0.6.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-brands-svg-icons": "^5.13.0",
@@ -83,7 +86,7 @@
     "@ngxs-labs/data": "^3.0.6",
     "@ngxs/logger-plugin": "^3.6.2",
     "@ngxs/store": "^3.6.2",
-    "@ts-stack/markdown": "^1",
+    "@ts-stack/markdown": "^1.2",
     "@types/escodegen": "0.0.6",
     "@types/eslint": "^7.2.0",
     "@types/esprima": "^4.0.2",
@@ -95,25 +98,23 @@
     "@typescript-eslint/eslint-plugin": "^3",
     "@typescript-eslint/parser": "^3",
     "angular2-uuid": "^1.1.1",
-    "eslint": "^6.8.0",
     "eslint-config-eslint": "^6.0.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-compat": "^3",
+    "eslint-plugin-compat": "^3.8",
     "eslint-plugin-jest": "^23.17.1",
     "eslint-plugin-lodash": "^7.1.0",
     "eslint-plugin-lodash-fp": "^2.2.0-a1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7",
+    "eslint-plugin-react": "^7.20.1",
     "eslint-plugin-react-hooks": "^4.0.4",
     "eslint-plugin-react-native": "^3.8.1",
     "eslint-plugin-react-redux": "^3.0.3",
     "eslint-plugin-sonarjs": "^0.5.0",
-    "eslint-plugin-unicorn": "^19",
+    "eslint-plugin-unicorn": "^20",
     "eslint-plugin-vue": "^6.2.2",
     "jest": "^26.1.0",
     "jest-junit": "^11.0.1",
     "jest-preset-angular": "^8.1.3",
-    "js-yaml": "^3.14.0",
     "ngx-contextmenu": "^5.4.0",
     "rxjs": "~6.5.4",
     "stylelint": "^13.3.3",
